adding build tag extended for prometheus/metrics (#2)
