+++ pkg/extensions/prometheus/metrics/metrics.go
@@ -1,3 +1,5 @@
+// +build extended
+
 package metrics
 
 import (
